// frontend/src/api/generated/models/inconsistencyDto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * MH
 * OpenAPI spec version: 1.0.0
 */
import type { InconsistencyDtoType } from './inconsistencyDtoType';

export interface InconsistencyDto {
  type: InconsistencyDtoType;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  variantName?: string | null;
  /** @nullable */
  mlItemId?: string | null;
  /** @nullable */
  mlVariationId?: string | null;
  /** Unidades en juego */
  units: number;
  /** @nullable */
  systemStock?: number | null;
  /** @nullable */
  mlStock?: number | null;
  /** @nullable */
  occurredAt?: string | null;
  detail: string;
}