// frontend/src/api/generated/models/inconsistencyDtoType.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * MH
 * OpenAPI spec version: 1.0.0
 */

export type InconsistencyDtoType = typeof InconsistencyDtoType[keyof typeof InconsistencyDtoType];


export const InconsistencyDtoType = {
  sale_without_stock: 'sale_without_stock',
  surplus_in_lots: 'surplus_in_lots',
  missing_in_lots: 'missing_in_lots',
} as const;